reconfiguring test folders and also making some updates
