+++ tests/integration/conftest.py
@@ -2,7 +2,6 @@ import os
 import re
 import uuid
 import time
-from collections import namedtuple
 from datetime import datetime
 from enum import Enum
 from random import randint
@@ -158,28 +157,6 @@ def rand_gen():
 @pytest.fixture
 def project(client, rand_gen):
     project = client.create_project(name=rand_gen(str))
-
-    def create_label(**kwargs):
-        """ Creates a label on a Legacy Editor project. Not supported in the new Editor.
-        Args:
-            **kwargs: Label attributes. At minimum, the label `DataRow`.
-        """
-        Label = Entity.Label
-        kwargs[Label.project] = project
-        kwargs[Label.seconds_to_label] = kwargs.get(Label.seconds_to_label.name,
-                                                    0.0)
-        data = {
-            Label.attribute(attr) if isinstance(attr, str) else attr:
-            value.uid if isinstance(value, DbObject) else value
-            for attr, value in kwargs.items()
-        }
-        query_str, params = query.create(Label, data)
-        query_str = query_str.replace(
-            "data: {", "data: {type: {connect: {name: \"Any\"}} ")
-        res = project.client.execute(query_str, params)
-        return Label(project.client, res["createLabel"])
-
-    project.create_label = create_label
     yield project
     project.delete()
 
@@ -203,21 +180,6 @@ def datarow(dataset, image_url):
     dr = next(dataset.data_rows())
     yield dr
     dr.delete()
-
-
-LabelPack = namedtuple("LabelPack", "project dataset data_row label")
-
-
-@pytest.fixture
-def label_pack(project, rand_gen, image_url):
-    client = project.client
-    dataset = client.create_dataset(name=rand_gen(str))
-    project.datasets.connect(dataset)
-    data_row = dataset.create_data_row(row_data=IMG_URL)
-    label = project.create_label(data_row=data_row, label=rand_gen(str))
-    time.sleep(10)
-    yield LabelPack(project, dataset, data_row, label)
-    dataset.delete()
 
 
 @pytest.fixture
@@ -299,6 +261,11 @@ def configured_project(project, client, rand_gen, image_url):
 
 @pytest.fixture
 def configured_project_with_label(client, rand_gen, image_url):
+    """Project with a connected dataset, having one datarow
+    Project contains an ontology with 1 bbox tool
+    Additionally includes a create_label method for any needed extra labels
+    One label is already created and yielded when using fixture
+    """
     project = client.create_project(name=rand_gen(str))
     dataset = client.create_dataset(name=rand_gen(str), projects=project)
     data_row = dataset.create_data_row(row_data=image_url)
@@ -310,7 +277,6 @@ def configured_project_with_label(client, rand_gen, image_url):
         Tool(tool=Tool.Type.BBOX, name="test-bbox-class"),
     ])
     project.setup(editor, ontology_builder.asdict())
-    project.enable_model_assisted_labeling()
     ontology = ontology_builder.from_project(project)
     predictions = [{
         "uuid": str(uuid.uuid4()),
@@ -325,13 +291,26 @@ def configured_project_with_label(client, rand_gen, image_url):
             "width": 50
         }
     }]
-    upload_task = LabelImport.create_from_objects(
-        client, project.uid, f'label-import-{uuid.uuid4()}', predictions)
-    upload_task.wait_until_done()
-    label = next(project.labels()).uid
-    yield [project, label]
+
+    def create_label():
+        """ Ad-hoc function to create a LabelImport
+        
+        Creates a LabelImport task which will create a label
+        """
+        upload_task = LabelImport.create_from_objects(
+            client, project.uid, f'label-import-{uuid.uuid4()}', predictions)
+        upload_task.wait_until_done(sleep_time_seconds=5)
+
+    project.create_label = create_label
+    project.create_label()
+    label = next(project.labels())
+    yield [project, dataset, data_row, label]
     dataset.delete()
     project.delete()
+    data_row.delete()
+
+    for label in project.labels():
+        label.delete()
 
 
 @pytest.fixture
