+++ tests/integration/test_toggle_mal.py
@@ -1,8 +1,3 @@
-from labelbox import Client
-from labelbox import Project
-import pytest
-
-
 def test_enable_model_assisted_labeling(project):
     response = project.enable_model_assisted_labeling()
     assert response == True
