+++ tests/unit/test_query.py
